+++ Files/SQL/StagingArea/CreateStageTableOrderLine_Stg.sql
@@ -8,9 +8,9 @@ IF NOT EXISTS (SELECT name FROM sys.tables WHERE name = 'OrderLine_Stg')
 	[SizeID] [int] NULL,
 	[LineTotal] [money] NULL,
 	[PromotionPercentage] [float] NULL,
-    [VATPercentage] [int] NULL,
+    [VATPercentage] [float] NULL,
 	[Currency] [nvarchar](3) NULL,
-	[SalePercentage] [int] NULL,
+	[SalePercentage] [float] NULL,
 	[Discount] [money] NULL,
     [SystemID] [nvarchar] (15) NULL,
 	[CreatedAt] [datetime] null
